requirements.txt
chainlit>=1.0.400
langchain>=0.1.16
langchain-openai>=0.0.6
langchain-community>=0.0.34
qdrant-client>=1.7.0
openai>=1.10.0,<2.0.0
python-dotenv>=1.0.0
pypdf>=3.17.1
python-docx>=1.0.1
unstructured>=0.10.30
python-magic>=0.4.27
python-magic-bin>=0.4.14
pdf2image>=1.16.3
pytesseract>=0.3.10
tiktoken>=0.5.2
